+++ javaparser-core/src/main/java/com/github/javaparser/JavaParser.java
@@ -22,13 +22,13 @@
 package com.github.javaparser;
 
 import com.github.javaparser.ast.CompilationUnit;
+import com.github.javaparser.ast.ImportDeclaration;
 import com.github.javaparser.ast.Node;
 import com.github.javaparser.ast.body.BodyDeclaration;
 import com.github.javaparser.ast.comments.CommentsCollection;
 import com.github.javaparser.ast.expr.AnnotationExpr;
 import com.github.javaparser.ast.expr.Expression;
 import com.github.javaparser.ast.expr.VariableDeclarationExpr;
-import com.github.javaparser.ast.imports.ImportDeclaration;
 import com.github.javaparser.ast.stmt.BlockStmt;
 import com.github.javaparser.ast.stmt.Statement;
 import com.github.javaparser.ast.type.ClassOrInterfaceType;
@@ -38,9 +38,7 @@ import java.nio.charset.Charset;
 import java.nio.file.Path;
 
 import static com.github.javaparser.ParseStart.*;
-import static com.github.javaparser.Providers.UTF8;
-import static com.github.javaparser.Providers.provider;
-import static com.github.javaparser.Providers.resourceProvider;
+import static com.github.javaparser.Providers.*;
 import static com.github.javaparser.utils.Utils.assertNotNull;
 
 /**
@@ -49,8 +47,6 @@ import static com.github.javaparser.utils.Utils.assertNotNull;
  * @author Júlio Vilmar Gesser
  */
 public final class JavaParser {
-    private static final JavaParser defaultInstance = new JavaParser();
-
     private final CommentsInserter commentsInserter;
     private final ParserConfiguration configuration;
 
@@ -105,8 +101,8 @@ public final class JavaParser {
                 commentsInserter.insertComments(resultNode, comments.copy().getComments());
             }
 
-            return new ParseResult<>(resultNode, parser.problems, parser.getTokens(),
-                    parser.getCommentsCollection());
+            return new ParseResult<>(resultNode, parser.problems, astParser.getTokens(),
+                    astParser.getCommentsCollection());
         } catch (Exception e) {
             return new ParseResult<>(e);
         } finally {
@@ -294,7 +290,7 @@ public final class JavaParser {
     }
 
     private static <T extends Node> T simplifiedParse(ParseStart<T> context, Provider provider) {
-        ParseResult<T> result = defaultInstance.parse(context, provider);
+        ParseResult<T> result = new JavaParser(new ParserConfiguration()).parse(context, provider);
         if (result.isSuccessful()) {
             return result.getResult().get();
         }