+++ javaparser-core/src/main/java/com/github/javaparser/ast/Node.java
@@ -26,6 +26,7 @@ import com.github.javaparser.Position;
 import com.github.javaparser.Range;
 import com.github.javaparser.ast.comments.BlockComment;
 import com.github.javaparser.ast.comments.Comment;
+import com.github.javaparser.ast.comments.JavadocComment;
 import com.github.javaparser.ast.comments.LineComment;
 import com.github.javaparser.ast.observer.AstObserver;
 import com.github.javaparser.ast.observer.ObservableProperty;
@@ -55,6 +56,7 @@ import static java.util.Collections.unmodifiableList;
  * <li>use a convenience method, like "addStatement(...)", or if none are available...</li>
  * <li>use a convenient constructor, like ClassOrInterfaceType(String name), or if none are available...</li>
  * <li>use the default constructor.</li>
+ * <li>Alternatively, use one of the JavaParser.parse(snippet) methods.</li>
  * </ul>
  * ... and use the various methods on the node to initialize it further, if needed.
  * <h2>Parent/child</h2>
@@ -215,6 +217,10 @@ public abstract class Node implements Cloneable, HasParentNode<Node>, Visitable 
         return this;
     }
 
+    public boolean hasJavaDocComment() {
+        return hasComment() && getComment() instanceof JavadocComment;
+    }
+
     /**
      * Use this to store additional information to this node.
      *
@@ -293,6 +299,14 @@ public abstract class Node implements Cloneable, HasParentNode<Node>, Visitable 
         comment.setParentNode(this);
     }
 
+    public boolean removeOrphanComment(Comment comment) {
+        boolean removed = orphanComments.remove(comment);
+        if (removed) {
+            comment.setParentNode(null);
+        }
+        return removed;
+    }
+
     /**
      * This is a list of Comment which are inside the node and are not associated
      * with any meaningful AST Node.
@@ -302,11 +316,13 @@ public abstract class Node implements Cloneable, HasParentNode<Node>, Visitable 
      * <p>
      * When more than one comment preceeds a statement, the one immediately preceding it
      * it is associated with the statements, while the others are orphans.
+     * <p>
+     * Changes to this list are not persisted.
      *
      * @return all comments that cannot be attributed to a concept
      */
     public List<Comment> getOrphanComments() {
-        return orphanComments;
+        return new LinkedList<>(orphanComments);
     }
 
     /**
@@ -376,10 +392,7 @@ public abstract class Node implements Cloneable, HasParentNode<Node>, Visitable 
     }
 
     public void tryAddImportToParentCompilationUnit(Class<?> clazz) {
-        CompilationUnit parentNode = getAncestorOfType(CompilationUnit.class);
-        if (parentNode != null) {
-            parentNode.addImport(clazz);
-        }
+        getAncestorOfType(CompilationUnit.class).ifPresent(p -> p.addImport(clazz));
     }
 
     /**
@@ -421,7 +434,6 @@ public abstract class Node implements Cloneable, HasParentNode<Node>, Visitable 
      * @param <M> The type of data
      * @param key The singleton key for the data
      * @param object The data object
-     * @throws IllegalArgumentException
      * @see DataKey
      */
     public <M> void setData(DataKey<M> key, M object) {