+++ javaparser-testing/src/test/java/com/github/javaparser/ast/NodeTest.java
@@ -22,6 +22,13 @@
 package com.github.javaparser.ast;
 
 import com.github.javaparser.JavaParser;
+import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
+import com.github.javaparser.ast.body.FieldDeclaration;
+import com.github.javaparser.ast.body.VariableDeclarator;
+import com.github.javaparser.ast.comments.Comment;
+import com.github.javaparser.ast.comments.BlockComment;
+import com.github.javaparser.ast.comments.JavadocComment;
+import com.github.javaparser.ast.comments.LineComment;
 import com.github.javaparser.ast.observer.AstObserver;
 import com.github.javaparser.ast.observer.AstObserverAdapter;
 import com.github.javaparser.ast.observer.ObservableProperty;
@@ -30,9 +37,12 @@ import org.junit.Test;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.EnumSet;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 public class NodeTest {
 
@@ -51,14 +61,14 @@ public class NodeTest {
 
         assertEquals(Arrays.asList(), changes);
 
-        cu.getClassByName("A").setName("MyCoolClass");
+        cu.getClassByName("A").get().setName("MyCoolClass");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
 
-        cu.getClassByName("MyCoolClass").getFieldByName("f").getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
+        cu.getClassByName("MyCoolClass").get().getFieldByName("f").get().getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
                 "VariableDeclarator.type changed from int to boolean"), changes);
 
-        cu.getClassByName("MyCoolClass").getMethodsByName("foo").get(0).getParamByName("p").setName("myParam");
+        cu.getClassByName("MyCoolClass").get().getMethodsByName("foo").get(0).getParameterByName("p").get().setName("myParam");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
                 "VariableDeclarator.type changed from int to boolean",
                 "Parameter.name changed from p to myParam"), changes);
@@ -75,20 +85,20 @@ public class NodeTest {
                 changes.add(String.format("%s.%s changed from %s to %s", observedNode.getClass().getSimpleName(), property.name().toLowerCase(), oldValue, newValue));
             }
         };
-        cu.getClassByName("A").register(observer, Node.ObserverRegistrationMode.JUST_THIS_NODE);
+        cu.getClassByName("A").get().register(observer, Node.ObserverRegistrationMode.JUST_THIS_NODE);
 
         assertEquals(Arrays.asList(), changes);
 
-        cu.getClassByName("A").setName("MyCoolClass");
+        cu.getClassByName("A").get().setName("MyCoolClass");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
 
-        cu.getClassByName("MyCoolClass").getFieldByName("f").getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
+        cu.getClassByName("MyCoolClass").get().getFieldByName("f").get().getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
 
-        cu.getClassByName("MyCoolClass").getMethodsByName("foo").get(0).getParamByName("p").setName("myParam");
+        cu.getClassByName("MyCoolClass").get().getMethodsByName("foo").get(0).getParameterByName("p").get().setName("myParam");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
 
-        cu.getClassByName("MyCoolClass").addField("int", "bar").getVariables().get(0).setInitializer("0");
+        cu.getClassByName("MyCoolClass").get().addField("int", "bar").getVariables().get(0).setInitializer("0");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
     }
 
@@ -103,23 +113,23 @@ public class NodeTest {
                 changes.add(String.format("%s.%s changed from %s to %s", observedNode.getClass().getSimpleName(), property.name().toLowerCase(), oldValue, newValue));
             }
         };
-        cu.getClassByName("A").register(observer, Node.ObserverRegistrationMode.THIS_NODE_AND_EXISTING_DESCENDANTS);
+        cu.getClassByName("A").get().register(observer, Node.ObserverRegistrationMode.THIS_NODE_AND_EXISTING_DESCENDANTS);
 
         assertEquals(Arrays.asList(), changes);
 
-        cu.getClassByName("A").setName("MyCoolClass");
+        cu.getClassByName("A").get().setName("MyCoolClass");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
 
-        cu.getClassByName("MyCoolClass").getFieldByName("f").getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
+        cu.getClassByName("MyCoolClass").get().getFieldByName("f").get().getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
                 "VariableDeclarator.type changed from int to boolean"), changes);
 
-        cu.getClassByName("MyCoolClass").getMethodsByName("foo").get(0).getParamByName("p").setName("myParam");
+        cu.getClassByName("MyCoolClass").get().getMethodsByName("foo").get(0).getParameterByName("p").get().setName("myParam");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
                 "VariableDeclarator.type changed from int to boolean",
                 "Parameter.name changed from p to myParam"), changes);
 
-        cu.getClassByName("MyCoolClass").addField("int", "bar").getVariables().get(0).setInitializer("0");
+        cu.getClassByName("MyCoolClass").get().addField("int", "bar").getVariables().get(0).setInitializer("0");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
                 "VariableDeclarator.type changed from int to boolean",
                 "Parameter.name changed from p to myParam"), changes);
@@ -136,23 +146,23 @@ public class NodeTest {
                 changes.add(String.format("%s.%s changed from %s to %s", observedNode.getClass().getSimpleName(), property.name().toLowerCase(), oldValue, newValue));
             }
         };
-        cu.getClassByName("A").register(observer, Node.ObserverRegistrationMode.SELF_PROPAGATING);
+        cu.getClassByName("A").get().register(observer, Node.ObserverRegistrationMode.SELF_PROPAGATING);
 
         assertEquals(Arrays.asList(), changes);
 
-        cu.getClassByName("A").setName("MyCoolClass");
+        cu.getClassByName("A").get().setName("MyCoolClass");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass"), changes);
 
-        cu.getClassByName("MyCoolClass").getFieldByName("f").getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
+        cu.getClassByName("MyCoolClass").get().getFieldByName("f").get().getVariable(0).setType(new PrimitiveType(PrimitiveType.Primitive.BOOLEAN));
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
                 "VariableDeclarator.type changed from int to boolean"), changes);
 
-        cu.getClassByName("MyCoolClass").getMethodsByName("foo").get(0).getParamByName("p").setName("myParam");
+        cu.getClassByName("MyCoolClass").get().getMethodsByName("foo").get(0).getParameterByName("p").get().setName("myParam");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
                 "VariableDeclarator.type changed from int to boolean",
                 "Parameter.name changed from p to myParam"), changes);
 
-        cu.getClassByName("MyCoolClass")
+        cu.getClassByName("MyCoolClass").get()
                 .addField("int", "bar")
                 .getVariables().get(0).setInitializer("0");
         assertEquals(Arrays.asList("ClassOrInterfaceDeclaration.name changed from A to MyCoolClass",
@@ -175,7 +185,7 @@ public class NodeTest {
         };
         cu.register(observer, Node.ObserverRegistrationMode.SELF_PROPAGATING);
 
-        cu.getClassByName("A").getMethodsByName("foo").get(0).getParameter(0).remove();
+        cu.getClassByName("A").get().getMethodsByName("foo").get(0).getParameter(0).remove();
         assertEquals(Arrays.asList("removing [int p] from index 0"), changes);
     }
 
@@ -194,7 +204,7 @@ public class NodeTest {
         cu.register(observer, Node.ObserverRegistrationMode.SELF_PROPAGATING);
 
         // I cannot remove the name of a type
-        assertEquals(false, cu.getClassByName("A").getName().remove());
+        assertEquals(false, cu.getClassByName("A").get().getName().remove());
         assertEquals(Arrays.asList(), changes);
     }
 
@@ -217,7 +227,71 @@ public class NodeTest {
         };
         cu.register(observer, Node.ObserverRegistrationMode.SELF_PROPAGATING);
 
-        assertEquals(true, cu.getClassByName("A").getMethodsByName("foo").get(0).getBody().get().remove());
+        assertEquals(true, cu.getClassByName("A").get().getMethodsByName("foo").get(0).getBody().get().remove());
         assertEquals(Arrays.asList("setting [BODY] to null"), changes);
+    }
+
+    @Test
+    public void removeOrphanCommentPositiveCase() {
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(EnumSet.noneOf(Modifier.class), false, "A");
+        Comment c = new LineComment("A comment");
+        decl.addOrphanComment(c);
+        assertEquals(1, decl.getOrphanComments().size());
+        assertTrue(decl == c.getParentNode().get());
+        assertTrue(decl.removeOrphanComment(c));
+        assertEquals(0, decl.getOrphanComments().size());
+        assertFalse(c.getParentNode().isPresent());
+    }
+
+    @Test
+    public void removeOrphanCommentNegativeCase() {
+        ClassOrInterfaceDeclaration aClass = new ClassOrInterfaceDeclaration(EnumSet.noneOf(Modifier.class), false, "A");
+        FieldDeclaration aField = new FieldDeclaration(EnumSet.noneOf(Modifier.class), new VariableDeclarator(PrimitiveType.intType(), "f"));
+        aClass.getMembers().add(aField);
+        Comment c = new LineComment("A comment");
+        aField.addOrphanComment(c);
+        // the comment is an orphan comment of the field, so trying to remove it on the class should not work
+        assertFalse(aClass.removeOrphanComment(c));
+        assertEquals(1, aField.getOrphanComments().size());
+        assertTrue(c.getParentNode().isPresent());
+    }
+
+    @Test
+    public void hasJavaDocCommentPositiveCaseWithSetJavaDocComment() {
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(EnumSet.noneOf(Modifier.class),
+                false, "Foo");
+        decl.setJavaDocComment("A comment");
+        assertEquals(true, decl.hasJavaDocComment());
+    }
+
+    @Test
+    public void hasJavaDocCommentPositiveCaseWithSetComment() {
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(EnumSet.noneOf(Modifier.class),
+                false, "Foo");
+        decl.setComment(new JavadocComment("A comment"));
+        assertEquals(true, decl.hasJavaDocComment());
+    }
+
+    @Test
+    public void hasJavaDocCommentNegativeCaseNoComment() {
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(EnumSet.noneOf(Modifier.class),
+                false, "Foo");
+        assertEquals(false, decl.hasJavaDocComment());
+    }
+
+    @Test
+    public void hasJavaDocCommentNegativeCaseLineComment() {
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(EnumSet.noneOf(Modifier.class),
+                false, "Foo");
+        decl.setComment(new LineComment("foo"));
+        assertEquals(false, decl.hasJavaDocComment());
+    }
+
+    @Test
+    public void hasJavaDocCommentNegativeCaseBlockComment() {
+        ClassOrInterfaceDeclaration decl = new ClassOrInterfaceDeclaration(EnumSet.noneOf(Modifier.class),
+                false, "Foo");
+        decl.setComment(new BlockComment("foo"));
+        assertEquals(false, decl.hasJavaDocComment());
     }
 }